+++ Data/TooltipData.lua
@@ -36,7 +36,12 @@ local function AddTooltipLine(tooltip, tooltipLine)
         return;
     end
 
-    local _, _, criteriaIsCompleted = GetAchievementCriteriaInfo(tooltipLine.AchievementId, tooltipLine.CriteriaIndex);
+    local criteriaIsCompleted;
+    if tooltipLine.CriteriaIndex ~= 0 then
+        _, _, criteriaIsCompleted = GetAchievementCriteriaInfo(tooltipLine.AchievementId, tooltipLine.CriteriaIndex);
+    else
+        criteriaIsCompleted = achievementIsCompleted;
+    end
     if criteriaIsCompleted and not addon.Options.db.Tooltip.Criteria.ShowIf.CriteriaIsCompleted then
         return;
     end
@@ -51,9 +56,13 @@ local function AddTooltipLine(tooltip, tooltipLine)
         text = tooltipLine.NotCompletedText;
         color = addon.Colors.RedRGB;
     end
-    text = text:ReplaceVars{
-        forAchievement = addon.Options.db.Tooltip.Criteria.ShowForAchievement and addon.L["for achievement"] or ""
-    };
+    if tooltipLine.CriteriaIndex ~= 0 then
+        text = text:ReplaceVars{
+            forAchievement = addon.Options.db.Tooltip.Criteria.ShowForAchievement and addon.L["for achievement"] or ""
+        };
+    else
+        text = name;
+    end
     tooltip:AddLine(icon .. " |T" .. achievementIcon .. ":0|t " .. string.trim(text:ReplaceVars{
         achievement = name
     }), color.R, color.G, color.B);
@@ -185,13 +194,30 @@ function tooltipData.Load()
     else
         TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.Unit, ProcessUnit100002);
         TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.Item, ProcessItem100002);
-        -- TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.Object, function(tooltip, localData)
-        --     addon.Diagnostics.DebugTable(localData);
+        -- for i = 0, 30, 1 do
+        --     TooltipDataProcessor.AddTooltipPostCall(i, function(tooltip, localData)
+        --         addon.Diagnostics.DebugTable(localData);
+        --         tooltip:AddLine(tostring(i));
+        --     end);
+        -- end
+        -- hooksecurefunc("WorldMap_GetQuestTimeForTooltip", function(questID)
+        --     print(questID);
+        -- end);
+        -- GameTooltip:HookScript("OnShow", function(self)
+        --     DebugTable = {};
+        --     local charactersPerLine = 1000;
+        --     for line in TableToString(self, charactersPerLine):gmatch("[^\r\n]+") do
+        --         tinsert(DebugTable, line);
+        --     end
         -- end);
     end
 
     data.ExportedTooltipData.Load(addon.Data.TooltipData);
 end
+
+-- function KrowiAF_DumpTooltip()
+--     DebugTable = GameTooltip;
+-- end
 
 -- local unitLink = "|cffffff00|Hunit:%s|h[%s]|h|r"