+++ Localization/enUS.PluginStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-01-19 17-25-20 ]] --
+-- [[ Exported at 2023-01-29 12-01-53 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --