+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 52.0 - 2023-02-02
+### Added
+- Trying to track/untrack an earned achievement will now untrack it (solution for completed tracked achievements)
+- The Earned By filter now has effect on the ability to track or untrack achievements (more logical behavior)
+- Feats of Strength filter
+- Option to only show events that have time data available (will prevent showing events saying 'No time data available')
+- Dragonrider Racing tooltip data
+
+### Changed
+- Split on login and on reload pop ups and chat messages options (settings might be corrupt and need to be reconfigured once)
+
+### Fixed
+- Earning a tracked achievement should now also untrack the achievement
+- Famliar Skies points is now 0 again instead of the previous 5
+
 ## 51.4 - 2023-01-25
 ### Added
 - Tooltip data for the The Winter Veil Gourmet achievement